+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/OnLoadResourceJS.java
@@ -11,7 +11,8 @@ public class OnLoadResourceJS {
           OnLoadResourceJS.ON_LOAD_RESOURCE_JS_SOURCE,
           UserScriptInjectionTime.AT_DOCUMENT_START,
           null,
-          false
+          false,
+          null
   );
 
   public static final String ON_LOAD_RESOURCE_JS_SOURCE = "window." + FLAG_VARIABLE_FOR_ON_LOAD_RESOURCE_JS_SOURCE + " = true;" +