+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/OnWindowBlurEventJS.java
@@ -10,7 +10,8 @@ public class OnWindowBlurEventJS {
           OnWindowBlurEventJS.ON_WINDOW_BLUR_EVENT_JS_SOURCE,
           UserScriptInjectionTime.AT_DOCUMENT_START,
           null,
-          false
+          false,
+          null
   );
 
   public static final String ON_WINDOW_BLUR_EVENT_JS_SOURCE = "(function(){" +