+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/OnWindowFocusEventJS.java
@@ -10,7 +10,8 @@ public class OnWindowFocusEventJS {
           OnWindowFocusEventJS.ON_WINDOW_FOCUS_EVENT_JS_SOURCE,
           UserScriptInjectionTime.AT_DOCUMENT_START,
           null,
-          false
+          false,
+          null
   );
 
   public static final String ON_WINDOW_FOCUS_EVENT_JS_SOURCE = "(function(){" +