+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/PrintJS.java
@@ -10,7 +10,8 @@ public class PrintJS {
           PrintJS.PRINT_JS_SOURCE,
           UserScriptInjectionTime.AT_DOCUMENT_START,
           null,
-          false
+          false,
+          null
   );
 
   public static final String PRINT_JS_SOURCE = "window.print = function() {" +