+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/plugin_scripts_js/PromisePolyfillJS.java
@@ -10,7 +10,8 @@ public class PromisePolyfillJS {
           PromisePolyfillJS.PROMISE_POLYFILL_JS_SOURCE,
           UserScriptInjectionTime.AT_DOCUMENT_START,
           null,
-          true
+          true,
+          null
   );
 
   // https://github.com/tildeio/rsvp.js