+++ lib/src/types/user_script.dart
@@ -3,6 +3,7 @@ import 'package:flutter_inappwebview_internal_annotations/flutter_inappwebview_i
 import '../in_app_webview/webview.dart';
 import 'user_script_injection_time.dart';
 import 'content_world.dart';
+import '../android/webview_feature.dart';
 
 part 'user_script.g.dart';
 
@@ -29,6 +30,11 @@ class UserScript_ {
   ///**NOTE**: available only on iOS.
   bool forMainFrameOnly;
 
+  ///A set of matching rules for the allowed origins.
+  ///
+  ///**NOTE**: available only on Android and only if [WebViewFeature.DOCUMENT_START_SCRIPT] feature is supported.
+  late Set<String> allowedOriginRules;
+
   ///A scope of execution in which to evaluate the script to prevent conflicts between different scripts.
   ///For more information about content worlds, see [ContentWorld].
   late ContentWorld contentWorld;
@@ -40,7 +46,10 @@ class UserScript_ {
         required this.injectionTime,
         @Deprecated("Use forMainFrameOnly instead") this.iosForMainFrameOnly,
         this.forMainFrameOnly = true,
+        Set<String>? allowedOriginRules,
         ContentWorld? contentWorld}) {
+    this.allowedOriginRules = allowedOriginRules != null ?
+        allowedOriginRules : Set.from(["*"]);
     this.contentWorld = contentWorld ?? ContentWorld.PAGE;
     // ignore: deprecated_member_use_from_same_package
     this.forMainFrameOnly = this.iosForMainFrameOnly != null