+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added `onCameraCaptureStateChanged`, `onMicrophoneCaptureStateChanged` WebView events
 - Added support for `onPermissionRequest` event on iOS 15.0+
 - Added `debugLoggingSettings` static property for WebView and ChromeSafariBrowser
+- Added `WebViewFeature.DOCUMENT_START_SCRIPT` Android feature support
 - Updated `getMetaThemeColor` on iOS 15.0+
 - Deprecated `onLoadError` for `onReceivedError`. `onReceivedError` will be called also for subframes
 - Deprecated `onLoadHttpError` for `onReceivedError`. `onReceivedHttpError` will be called also for subframes